✨ 검색 페이지 리팩토링: 검색 로직을 커스텀 훅으로 분리하고, UI 구성 요소를 모듈화하여 코드 가독성 및 유지보수성 향상. 필터 및 검색 결과 컴포넌트 추가, API 호출 로직 통합.
